change the lib desc: crawling -> extracting.
With the suggest of Tyler Xing (@CNBorn), this commit changed the
library description by using "extracting" instead of "crawling".

This library could provide some feature about building extracting
pipeline for web data. But it is not and will not be a alternative of
other **crawling** framework such as scrapy. You may use scrapy to
extract web page before, then now you can migrate to us. But if you used
scrapy to build a crawling spider, we can not replace it.